Changed decimal to double (because in database is double)

+++ CleanArchitectureUsersApp/CleanArchitectureUsersApp.Application/DTOs/Responses/UserResponse.cs
@@ -8,8 +8,8 @@
         public string Email { get; set; }
         public string AddressStreet { get; set; }
         public string AddressCity { get; set; }
-        public decimal GeoLat { get; set; }
-        public decimal GeoLng { get; set; }
+        public double GeoLat { get; set; }
+        public double GeoLng { get; set; }
         public string? Website { get; set; }
         public bool IsActive { get; set; }
     }
